Correcoes e ajustes.

+++ shared/Campaign.Shared/DataBaseContext/Entities/Order/OrderDetail.cs
@@ -3,8 +3,8 @@
     public class OrderDetail
     {
         public int Id { get; private set; }
-        public long Quantity { get; private set; }
-        public int Price { get; private set; }
+        public decimal Quantity { get; private set; }
+        public decimal Price { get; private set; }
         public int CustomerId { get; private set; }
         public int ProductId { get; private set; }
     }
